pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "network-tools"
version = "0.1.0"
description = "Automated network device discovery and inventory management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "RA"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
scan = [
    "scapy>=2.5",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.11",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.5",
    "types-PyYAML>=6.0",
]

[project.scripts]
network-tools = "network_tools.cli.main:cli"

[project.urls]
"Homepage" = "https://github.com/score-ra/network-tools"
"Bug Tracker" = "https://github.com/score-ra/network-tools/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
